Better handling of 404 routes (#862)
