+++ lib/handleRequest.js
@@ -46,7 +46,8 @@ function handleRequest (req, res, params, context) {
     return
   }
 
-  reply.code(405).send(new Error('Method Not Allowed: ' + method))
+  // Return 404 instead of 405 see https://github.com/fastify/fastify/pull/862 for discussion
+  reply.code(404).send(new Error('Not Found'))
 }
 
 function handler (reply) {
